+++ src/main/java/org/jabref/gui/importer/ImportEntriesDialog.java
@@ -1,6 +1,7 @@
 package org.jabref.gui.importer;
 
 import java.util.EnumSet;
+import java.util.Optional;
 
 import javax.swing.undo.UndoManager;
 
@@ -12,6 +13,7 @@ import javafx.scene.Node;
 import javafx.scene.control.Button;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.CheckBox;
+import javafx.scene.control.ComboBox;
 import javafx.scene.control.Label;
 import javafx.scene.control.ToggleButton;
 import javafx.scene.control.Tooltip;
@@ -21,7 +23,6 @@ import javafx.scene.layout.VBox;
 import javafx.scene.text.Text;
 
 import org.jabref.gui.DialogService;
-import org.jabref.gui.Globals;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.util.BackgroundTask;
@@ -32,6 +33,8 @@ import org.jabref.gui.util.TextFlowLimited;
 import org.jabref.gui.util.ViewModelListCellFactory;
 import org.jabref.logic.importer.ParserResult;
 import org.jabref.logic.l10n.Localization;
+import org.jabref.logic.shared.DatabaseLocation;
+import org.jabref.logic.util.io.FileUtil;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.BibEntryTypesManager;
@@ -49,6 +52,7 @@ import org.controlsfx.control.CheckListView;
 public class ImportEntriesDialog extends BaseDialog<Boolean> {
 
     public CheckListView<BibEntry> entriesListView;
+    public ComboBox<BibDatabaseContext> libraryListView;
     public ButtonType importButton;
     public Label totalItems;
     public Label selectedItems;
@@ -102,6 +106,28 @@ public class ImportEntriesDialog extends BaseDialog<Boolean> {
         entriesListView.setPlaceholder(placeholder);
         entriesListView.setItems(viewModel.getEntries());
 
+        libraryListView.setEditable(false);
+        libraryListView.getItems().addAll(stateManager.getOpenDatabases());
+        new ViewModelListCellFactory<BibDatabaseContext>()
+                .withText(database -> {
+                    Optional<String> dbOpt = Optional.empty();
+                    if (database.getDatabasePath().isPresent()) {
+                        dbOpt = FileUtil.getUniquePathFragment(stateManager.collectAllDatabasePaths(), database.getDatabasePath().get());
+                    }
+                    if (database.getLocation() == DatabaseLocation.SHARED) {
+                        return database.getDBMSSynchronizer().getDBName() + " [" + Localization.lang("shared") + "]";
+                    }
+
+                    if (dbOpt.isEmpty()) {
+                        return Localization.lang("untitled");
+                    }
+
+                    return dbOpt.get();
+                })
+                .install(libraryListView);
+        viewModel.selectedDbProperty().bind(libraryListView.getSelectionModel().selectedItemProperty());
+        stateManager.getActiveDatabase().ifPresent(database1 -> libraryListView.getSelectionModel().select(database1));
+
         PseudoClass entrySelected = PseudoClass.getPseudoClass("entry-selected");
         new ViewModelListCellFactory<BibEntry>()
                 .withGraphic(entry -> {
@@ -129,7 +155,7 @@ public class ImportEntriesDialog extends BaseDialog<Boolean> {
                             duplicateButton.setOnAction(event -> viewModel.resolveDuplicate(entry));
                             container.getChildren().add(1, duplicateButton);
                         }
-                    }).executeWith(Globals.TASK_EXECUTOR);
+                    }).executeWith(taskExecutor);
 
                     /*
                     inserted the if-statement here, since a Platform.runLater() call did not work.